+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Date format: DD/MM/YYYY
 - TreeViewItem now has custom primary key (`value` field)
 - Added `onSelectionChanged` callback to `TreeView`
 - Account for enabled on pressing states ([#233](https://github.com/bdlukaa/fluent_ui/pull/233))
+- Implement `CommandBar` with dynamic overflow ([#232](https://github.com/bdlukaa/fluent_ui/pull/232))
+- Add `DynamicOverflow` layout widget, for one-run horizontal or vertical layout with an overflow widget
+- Add `HorizontalScrollView` helper widget, with mouse wheel horizontal scrolling
 
 ## [3.9.1] - Input Update - [25/02/2022]